requirements.txt
# Core
numpy
pandas
scikit-learn
pyyaml
tqdm

# PyTorch
torch
transformers

# Vector DB (choose one; FAISS is easiest)
faiss-cpu

# App
streamlit

# Optional quality tools (recommended)
ruff
black
